+++ src/canvas/dialogs/help_dialog.rs
@@ -22,7 +22,7 @@ impl HelpDialog for Painter {
         &self, f: &mut Frame<'_, B>, app_state: &mut App, draw_loc: Rect,
     ) {
         let repeat_num = usize::from(draw_loc.width).saturating_sub(HELP_BASE.chars().count() + 2);
-        let help_title = format!(" Help ─{}─ Esc to close ", "─".repeat(repeat_num as usize));
+        let help_title = format!(" Help ─{}─ Esc to close ", "─".repeat(repeat_num));
 
         if app_state.is_force_redraw {
             // We must also recalculate how many lines are wrapping to properly get scrolling to work on
