+++ src/canvas/widgets/battery_display.rs
@@ -31,10 +31,7 @@ impl BatteryDisplayWidget for Painter {
                 const TITLE_BASE: &str = " Battery ── Esc to go back ";
                 let repeat_num =
                     usize::from(draw_loc.width).saturating_sub(TITLE_BASE.chars().count() + 2);
-                let result_title = format!(
-                    " Battery ─{}─ Esc to go back ",
-                    "─".repeat(repeat_num as usize)
-                );
+                let result_title = format!(" Battery ─{}─ Esc to go back ", "─".repeat(repeat_num));
                 result_title
             } else {
                 " Battery ".to_string()
